fix: remove store and useless packages.
