+++ package.json
@@ -66,15 +66,11 @@
     "@types/node": "^11.13.0",
     "@types/react": "^16.8.12",
     "@types/react-dom": "^16.8.3",
-    "@types/react-redux": "^7.0.6",
     "@types/react-router-dom": "^4.3.1",
     "classnames": "^2.2.6",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
-    "react-redux": "^6.0.1",
     "react-router-dom": "^5.0.0",
-    "redux": "^4.0.1",
-    "redux-saga": "^1.0.2",
     "uninstall": "0.0.0"
   }
 }
