--- src/store/user/model.ts
@@ -1,3 +1,0 @@
-export interface UserInfo {
-  fullname: string | null;
-}
